native, native4d: exact int64 Reshape, Permute, arithmetic and casts on Direct
Eval_direct routes an I64 operand of Reshape, Permute, tensor-tensor
Add/Sub/Mul, Mul_scalar and To_copy through Compute_i64 functors that read via
Tensor.read_i64_at6, so values past 2^53 no longer round-trip through
Payload.get_float. To_copy(Long) casts a float with the checked Float-to-I64
policy, the builders thread the I64 format onto an output edge only when the
operands are I64, and a mixed I64/F32 pair is rejected at checked admission.
Native4D mirrors the same arms, and bridge fixtures compare squeeze, unsqueeze
and the arithmetic with real ATen.

End-to-end fixtures chain mvitv2's and EdgeNeXt's real node sequences through
Graph_builder and Eval_direct.run, so a wiring mistake between two correct arms
shows up.

+++ test/native_interp/to_copy_test.ml
@@ -62,8 +62,8 @@ let%expect_test "_to_copy dtype=FLOAT/LONG/BOOL lowers to the matching target" =
     graph
     inputs: [t0 f32 [C=3] ->[n0]]
     nodes:
-      n0: [t1 f32 [C=3]] = to_copy x=t0 target=long
-    outputs: [t1 f32 [C=3] <-n0]
+      n0: [t1 i64 [C=3]] = to_copy x=t0 target=long
+    outputs: [t1 i64 [C=3] <-n0]
     BOOL
     graph
     inputs: [t0 f32 [C=3] ->[n0]]
@@ -92,11 +92,11 @@ let%expect_test "_to_copy: non_blocking carries no computational effect" =
     graph
     inputs: [t0 f32 [C=3] ->[n0]]
     nodes:
-      n0: [t1 f32 [C=3]] = to_copy x=t0 target=long
-    outputs: [t1 f32 [C=3] <-n0]
+      n0: [t1 i64 [C=3]] = to_copy x=t0 target=long
+    outputs: [t1 i64 [C=3] <-n0]
     non_blocking=true
     graph
     inputs: [t0 f32 [C=3] ->[n0]]
     nodes:
-      n0: [t1 f32 [C=3]] = to_copy x=t0 target=long
-    outputs: [t1 f32 [C=3] <-n0] |}]
+      n0: [t1 i64 [C=3]] = to_copy x=t0 target=long
+    outputs: [t1 i64 [C=3] <-n0] |}]
